feat(frontend): ExtensionActivity rendering via shared CefCanvas
Pulls canvas + worker + input handling out of BrowserActivity into
CefCanvas. ExtensionActivity uses it without a Toolbar. PaneContent
extension branch now renders ExtensionActivity instead of an iframe.
Cleanup of useIframeKeydownBridge / iframe_url field / handlers_ws SDK
is deferred to Task 13.
