+++ daemon/frontend/src/extension/index.ts
@@ -1,0 +1,2 @@
+//! Module barrel for the CEF-backed ExtensionActivity.
+export { ExtensionActivity } from './ExtensionActivity';
