+++ daemon/frontend/src/layout/PaneContent.tsx
@@ -1,11 +1,10 @@
-import { useEffect, useRef } from 'react';
 import { BrowserActivity } from '../browser';
+import { ExtensionActivity } from '../extension';
 import { Terminal } from '../terminal/Terminal';
 import { ClickShield } from './ClickShield';
 import { PanePlaceholder } from './PanePlaceholder';
 import { PaneTabBar } from './PaneTabBar';
 import type { PaneView } from './types';
-import { useIframeKeydownBridge } from './useIframeKeydownBridge';
 
 interface PaneContentProps {
   windowId: string;
@@ -30,14 +29,11 @@ function PaneBody({ windowId, pane, isActive, onActivate }: PaneContentProps) {
   if (!activity) return <PanePlaceholder paneId={pane.id} />;
 
   if (activity.kind === 'extension') {
-    if (!activity.iframe_url) return <PanePlaceholder paneId={pane.id} />;
     return (
-      <IframePane
-        url={activity.iframe_url}
-        title={`extension-${activity.id}`}
-        isActive={isActive}
-        onActivate={onActivate}
-      />
+      <>
+        <ExtensionActivity windowId={windowId} paneId={pane.id} activityId={activity.id} />
+        {!isActive && <ClickShield onActivate={onActivate} />}
+      </>
     );
   }
   if (activity.kind === 'browser') {
@@ -56,32 +52,5 @@ function PaneBody({ windowId, pane, isActive, onActivate }: PaneContentProps) {
       activityId={activity.id}
       isActive={isActive}
     />
-  );
-}
-
-interface IframePaneProps {
-  url: string;
-  title: string;
-  isActive: boolean;
-  onActivate: () => void;
-}
-
-function IframePane({ url, title, isActive, onActivate }: IframePaneProps) {
-  const iframeRef = useRef<HTMLIFrameElement>(null);
-  const prevActiveRef = useRef(isActive);
-  useEffect(() => {
-    if (isActive && !prevActiveRef.current) {
-      iframeRef.current?.contentWindow?.focus();
-    }
-    prevActiveRef.current = isActive;
-  }, [isActive]);
-
-  useIframeKeydownBridge(iframeRef);
-
-  return (
-    <>
-      <iframe ref={iframeRef} src={url} title={title} className="h-full w-full border-0" />
-      {!isActive && <ClickShield onActivate={onActivate} />}
-    </>
   );
 }
